Meetup Screen working functionally. Add details and full UI next.

+++ src/utilities/cancel-invite.js
@@ -1,13 +1,17 @@
 import firestore from '@react-native-firebase/firestore'
 
 export default async invite => {
-  try {
-    await firestore()
-      .collection('meetups')
-      .doc(invite.id)
-      .delete()
-    return invite
-  } catch (err) {
-    throw err
+  if (invite.sender.uid === global.user.uid) {
+    try {
+      await firestore()
+        .collection('meetups')
+        .doc(invite.id)
+        .delete()
+      return invite
+    } catch (err) {
+      throw err
+    }
+  } else {
+    throw Error("Cannot delete an invite you didn't create.")
   }
 }
